requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
fastapi==0.112.0
pdfminer.six==20240706
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
scikit-learn==1.5.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
uvicorn==0.30.5